refactor(testing): move test koin modules to specific directories
