+++ core/database/src/desktopTest/kotlin/org/mifos/core/database/di/TestDatabaseModule.desktop.kt
@@ -7,7 +7,7 @@
  *
  * See See https://github.com/openMF/kmp-project-template/blob/main/LICENSE
  */
-package org.mifos.testing.di
+package org.mifos.core.database.di
 
 import androidx.room.Room
 import androidx.sqlite.driver.bundled.BundledSQLiteDriver
